Move all hashable tests to same directory
