+++ tests/hashable/chunk_strategies.py
