+++ tests/hashable/hashable_strategies.py
