+++ tests/hashable/test_chunk_updates.py
@@ -8,7 +8,7 @@ from ssz.hashable_structure import (
     update_element_in_chunk,
     update_elements_in_chunk,
 )
-from tests.tree_hash.strategies import (
+from tests.hashable.chunk_strategies import (
     chunk_st,
     chunk_updates_st,
     element_size_st,
